Leverage some assertions related to inotify events.
As testing showed, in real life things happen concurrently, so
despite the synchronous nature of folly/nully itself, events may
get handled with a delay, their order might change, etc, therefore
we cannot expect that events and states adhere to a simple linear logic.
For this reason, here we drop those assertions which tried to enforce
such a logic. There is no problem with proceeding on after the
failure of these assertions, we just get some false alarms.